refactor: fix chart error & move form components to seperate folders
