+++ client/src/components/ui/form/session/SessionGearForm.vue
@@ -29,9 +29,9 @@
 <script lang="ts">
 import Vue from 'vue'
 
-import helper from '../../../services/helper'
+import helper from '@/services/helper'
 
-import { Session } from '../../../../../shared/interfaces/Session'
+import { Session } from '../../../../../../shared/interfaces'
 
 export default Vue.extend({
   name: 'SessionGearForm',
