+++ shared/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './Gear'
+export * from './spot'
+export * from './Session'
+export * from './User'
